msf-worker: do not add route to the own network
Do not add a route to the network the attacker is in through the
established session or do an arp scan in it. This reduces the scan
time. Someday later this is needed again if we want to attack some
devices like which are directly accessible by the attacker, but are
blocking his IP.